feat: 多项优化
1. 切换页面增加动画过渡
2. 软件管理默认显示第三方应用
3. 解决软件管理闪退问题
4. 解决文件传输闪退问题
5. 增加部分软件操作后自动刷新软件列表
6. 修复启动刷机工具启动boot,刷写和清楚不可用
7. 修复刷机工具点击重启按钮和i闪退
8. 修复设备信息界面的设备控制不可用
9. 优化设备列表的显示

+++ src/cpp/flashPageTools/fastbootdevicemanager.cpp
@@ -1,6 +1,7 @@
 #include "fastbootdevicemanager.h"
 #include <QDebug>
 #include "../utils/utils.hpp"
+#include "src/cpp/utils/Notification.h"
 
 DeviceCheckTool::DeviceCheckTool(QObject *parent)
     : QObject (parent)
@@ -94,11 +95,15 @@ void FastBootDeviceManager::rebootToFastBoot(const QString &deviceCode)
     if (deviceCode.isEmpty()) {
         tarDevice = m_currentDeviceCode;
     }
+    if (m_currentDeviceCode.isEmpty()) {
+        NotificationControl::instance()->send("当前无设备连接", NotificationControl::Warning, 3000);
+        return;
+    }
 
     QStringList args;
     args << "-s"  << tarDevice << "reboot" << "fastboot";
 
-    auto operatorFunc = [&tarDevice, &args](){
+    auto operatorFunc = [tarDevice, args](){
         ADBTools::instance()->executeCommand(ADBTools::FASTBOOT, args);
     };
 
@@ -111,11 +116,15 @@ void FastBootDeviceManager::rebootToSystem(const QString &deviceCode)
     if (deviceCode.isEmpty()) {
         tarDevice = m_currentDeviceCode;
     }
+    if (m_currentDeviceCode.isEmpty()) {
+        NotificationControl::instance()->send("当前无设备连接", NotificationControl::Warning, 3000);
+        return;
+    }
 
     QStringList args;
     args << "-s" << tarDevice << "reboot";
 
-    auto operatorFunc = [&tarDevice, &args](){
+    auto operatorFunc = [tarDevice, args](){
         ADBTools::instance()->executeCommand(ADBTools::FASTBOOT, args);
     };
 
@@ -128,11 +137,15 @@ void FastBootDeviceManager::rebootToRecovery(const QString &deviceCode)
     if (deviceCode.isEmpty()) {
         tarDevice = m_currentDeviceCode;
     }
+    if (m_currentDeviceCode.isEmpty()) {
+        NotificationControl::instance()->send("当前无设备连接", NotificationControl::Warning, 3000);
+        return;
+    }
 
     QStringList args;
     args << "-s" << tarDevice << "reboot" << "recovery";
 
-    auto operatorFunc = [&tarDevice, &args](){
+    auto operatorFunc = [tarDevice, args](){
         ADBTools::instance()->executeCommand(ADBTools::FASTBOOT, args);
     };
 
@@ -145,12 +158,16 @@ void FastBootDeviceManager::powerOff(const QString &deviceCode)
     if (deviceCode.isEmpty()) {
         tarDevice = m_currentDeviceCode;
     }
+    if (m_currentDeviceCode.isEmpty()) {
+        NotificationControl::instance()->send("当前无设备连接", NotificationControl::Warning, 3000);
+        return;
+    }
 
     QStringList args;
     args << "-s" << tarDevice << "poweroff";
 
 
-    auto operatorFunc = [&tarDevice, &args](){
+    auto operatorFunc = [tarDevice, args](){
         ADBTools::instance()->executeCommand(ADBTools::FASTBOOT, args);
     };
